+++ common/src/commonMain/kotlin/fr/outadoc/aoc/twentytwenty/Day02.kt
@@ -28,22 +28,22 @@ class Day02 : Day(Year.TwentyTwenty) {
         }
     }
 
-    fun step1(): Long {
+    fun step1(): Int {
         fun isValid(entry: PasswordEntry): Boolean {
             val letterCount = entry.password.count { c -> c == entry.policy.letter }
             return letterCount in entry.policy.first..entry.policy.second
         }
 
-        return input.count { entry -> isValid(entry) }.toLong()
+        return input.count { entry -> isValid(entry) }
     }
 
-    fun step2(): Long {
+    fun step2(): Int {
         fun isValid(entry: PasswordEntry): Boolean {
             val isFirstLetterOk = entry.password[entry.policy.first - 1] == entry.policy.letter
             val isSecondLetterOk = entry.password[entry.policy.second - 1] == entry.policy.letter
             return isFirstLetterOk xor isSecondLetterOk
         }
 
-        return input.count { entry -> isValid(entry) }.toLong()
+        return input.count { entry -> isValid(entry) }
     }
 }
